+++ src/Button/Group.php
@@ -2,23 +2,23 @@
 /**
  *	...
  *	@category		Library
- *	@package		CeusMedia_Bootstrap
+ *	@package		CeusMedia_Bootstrap_Button
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@copyright		2012-2015 {@link http://ceusmedia.de/ Ceus Media}
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Bootstrap
  */
-namespace CeusMedia\Bootstrap;
+namespace CeusMedia\Bootstrap\Button;
 /**
  *	...
  *	@category		Library
- *	@package		CeusMedia_Bootstrap
+ *	@package		CeusMedia_Bootstrap_Button
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@copyright		2012-2015 {@link http://ceusmedia.de/ Ceus Media}
  *	@license		http://www.gnu.org/licenses/gpl-3.0.txt GPL 3
  *	@link			https://github.com/CeusMedia/Bootstrap
  */
-class ButtonGroup{
+class Group{
 	protected $buttons		= array();
 	protected $stacked		= FALSE;
 	protected $class		= "";
